// lib/generated/prisma/models/Place.ts
export type PlaceModel = runtime.Types.Result.DefaultSelection<Prisma.$PlacePayload>
// ...
export type $PlacePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Place"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    placeId: string
    name: string
    rating: number | null
    totalRatings: number | null
    priceLevel: number | null
    address: string | null
    openNow: boolean | null
    photoUrl: string | null
    lat: number | null
    lng: number | null
    types: string
    searchAliases: string
    lastSyncedAt: Date
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["place"]>
  composites: {}
}